% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neocities_info}
\alias{neocities_info}
\title{Get information on a Neocities site.}
\usage{
neocities_info(user = neocities_user(), ...)
}
\arguments{
\item{user}{The username for your neocities account. See ?neocities_user for more details.}

\item{...}{Arguments that will be passed to the underlying httr GET request.}
}
\description{
Get information on a Neocities site.
}
\examples{
\dontrun{

## Username has not been set as an environmental variable, so
## the user will be prompted to enter their username if in
## interactive mode.

neocities_info()

## Setting the username as an environmental variable is wise when not
## in interactive mode.

Sys.setenv(NEOCITIES_USERNAME = "username")
neocities_info()

## You can also pass your username as a function argument.

neocities_info(user="username")

}
}

