% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neocities_pass}
\alias{neocities_pass}
\title{Interactively set your neocities password.}
\usage{
neocities_pass(force = FALSE)
}
\arguments{
\item{force}{If set to \code{TRUE}, this function will attempt to return the neocities password variable, even if it does not exist. The default value is \code{FALSE}.}
}
\description{
Interactively set your neocities password.
}

