% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neocities_upload}
\alias{neocities_upload}
\title{Upload a file to your Neocities site.}
\usage{
neocities_upload(path, ..., user = neocities_user(),
  pass = neocities_pass(), filename = NULL)
}
\arguments{
\item{path}{The path to the file you want to upload.}

\item{...}{Arguments that will be passed to the underlying httr POST request.}

\item{user}{The username for your neocities account. See ?neocities_user for more details.}

\item{pass}{The password for your neocities account. See ?neocities_pass for more details.}

\item{filename}{The name of the file as it will appear on neocities.org. If \code{NULL}, the base name of the file will be used. The default value is \code{NULL}.}
}
\description{
Upload a file to your Neocities site.
}
\examples{
\dontrun{

## Username and password have not been set as environmental variables, so
## the user will be prompted to enter their username and password if in
## interactive mode. This file will appear on neocities.org at
## http://username.neocities.org/report.html

neocities_upload("~/Desktop/report.html")

## Setting username and password as environmental variables is wise when not
## in interactive mode, but be sure not to distribute them with your code.

Sys.setenv(NEOCITIES_USERNAME = "username")
Sys.setenv(NEOCITIES_PASSWORD = "password")
neocities_upload("~/Desktop/report.html")

## You can also pass your username and password as function arguments.

neocities_upload("~/Desktop/report.html", user="username", pass="password")

## You can change what the filename will be on your neocities domain. This
## file will appear at http://username.neocities.org/october.html

neocities_upload("~/Desktop/report.html", filename="october.html")

}
}

