% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neocities_user}
\alias{neocities_user}
\title{Interactively set your neocities username.}
\usage{
neocities_user(force = FALSE)
}
\arguments{
\item{force}{If set to \code{TRUE}, this function will attempt to return the neocities username variable, even if it does not exist. The default value is \code{FALSE}.}
}
\description{
Interactively set your neocities username.
}

